atproto_firehose.subscribe: let lexrpc Client parse CAR header/payload
I think at one point we moved the parsing here so that we could skip parsing the payload for event types we don't care about, but those are rare, and instead we end up re-parsing the header once or twice. this lets us narrow down our explicit usage of libipld in atproto_firehose. re #1316